# src/components/docs/docSwagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is WEBSTORE API REQUIREMENT OF TRONTRON",
    "version": "1.0.5",
    "title": "Swagger WebStore",
    "contact": {
      "email": "lenguyenyenhong1199@gmail.com"
    }
  },
  "host": "http://localhost:5000",
  "basePath":"/",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/staff/":{
      "post": {
        "tags": [
          "Staff"
        ],
        "summary": "Add new Staff",
        "description": "",
        "operationId": "createNewStaff",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Staff"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"200"
                },
                "success":{
                  "type":"boolean",
                  "example":"true"
                },
                "data":{
                  
                    "$ref": "#/definitions/Staff"             
                  
                }
              }
            }
          },
          "400":{
            "description":"Error Validate",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"400"
                },
                "success":{
                  "type":"boolean",
                  "example":"false"
                },
                "data":{
                  "type": "string",
                  "example":"Validate Error"
                  }
                }
              }
            },
         "500":{
          "description":"Server error",
          "schema": {
            "type":"object",
            "properties":{
              "code": {
                "type":"integer",
                "example":"500"
              },
              "success":{
                "type":"boolean",
                "example":"false"
              },
              "data":{
                "type": "string",
                "example":"Server error"
                }
              }
            }
         }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
    "/staff/{staffId}":{
      "get": {
        "tags": [
          "Staff"
        ],
        "summary": "Find Staff by ID",
        "description": "Returns a single Staff",
        "operationId": "getStaffById",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "staffId",
            "in": "path",
            "description": "ID of Staff to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"200"
                },
                "success":{
                  "type":"boolean",
                  "example":"true"
                },
                "data":{
                  
                    "$ref": "#/definitions/Staff"             
                  
                }
              }
            }
          },
          "400":{
            "description":"Error Validate",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"400"
                },
                "success":{
                  "type":"boolean",
                  "example":"false"
                },
                "data":{
                  "type": "string",
                  "example":"Validate Error"
                  }
                }
              }
            },
         "500":{
          "description":"Server error",
          "schema": {
            "type":"object",
            "properties":{
              "code": {
                "type":"integer",
                "example":"500"
              },
              "success":{
                "type":"boolean",
                "example":"false"
              },
              "data":{
                "type": "string",
                "example":"Server error"
                }
              }
            }
         }
        },
        "security": [
          {
            "api_key": [
              
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "Staff"
        ],
        "summary": "Updates an Staff in the store with form data",
        "description": "",
        "operationId": "updateStaffById",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "staffId",
            "in": "path",
            "description": "ID of staff that needs to be updated",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "staffName",
            "in": "body",
            "description": "Updated name of the Staff",
            "required": false,
            "type": "string"
          },
          {
            "name": "email",
            "in": "body",
            "description": "Updated email of the Staff",
            "required": false,
            "type": "string",
            "format":"email"
          },
          {
            "name": "address",
            "in": "body",
            "description": "Updated address of the Staff",
            "required": false,
            "type": "string"
          },
          {
            "name": "phone",
            "in": "body",
            "description": "Updated phone of the Staff",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"200"
                },
                "success":{
                  "type":"boolean",
                  "example":"true"
                },
                "data":{
                  
                    "$ref": "#/definitions/Staff"             
                  
                }
              }
            }
          },
          "400":{
            "description":"Error Validate",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"400"
                },
                "success":{
                  "type":"boolean",
                  "example":"false"
                },
                "data":{
                  "type": "string",
                  "example":"Validate Error"
                  }
                }
              }
            },
         "500":{
          "description":"Server error",
          "schema": {
            "type":"object",
            "properties":{
              "code": {
                "type":"integer",
                "example":"500"
              },
              "success":{
                "type":"boolean",
                "example":"false"
              },
              "data":{
                "type": "string",
                "example":"Server error"
                }
              }
            }
         }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Staff"
        ],
        "summary": "Deletes a Staff",
        "description": "Delete Stafff By Id Staff",
        "operationId": "deleteStaffById",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "staffId",
            "in": "path",
            "description": "staffId to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"200"
                },
                "success":{
                  "type":"boolean",
                  "example":"true"
                },
                "data":{
                  "type": "string",
                  "example":"Delete Successfully"                 
                }
              }
            }
          },
          "400":{
            "description":"Error Validate",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"400"
                },
                "success":{
                  "type":"boolean",
                  "example":"false"
                },
                "data":{
                  "type": "string",
                  "example":"Validate Error"
                  }
                }
              }
            },
         "500":{
          "description":"Server error",
          "schema": {
            "type":"object",
            "properties":{
              "code": {
                "type":"integer",
                "example":"500"
              },
              "success":{
                "type":"boolean",
                "example":"false"
              },
              "data":{
                "type": "string",
                "example":"Server error"
                }
              }
            }
         }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read:pets": "read your pets",
        "write:pets": "modify pets in your account"
      }
    }
  },
  "definitions": {
...
    "Staff": {
      "type": "object",
      "properties": {
        "staffName": {
          "type": "string"
        },
        "dateOfBirth": {
          "type": "string",
          "format": "date-time"
        },
        "address": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "isActive": {
          "type": "boolean",        
          "description": "Staff Status"
        }
      }
    },
...
  },
...
}